Fix errors with helpers, handlers, and containers to compile
There were lots of errors throughout the project which wouldn't let it
compile as well as some naming convention issues that would be useful to
fix now before we get too far.

I moved `containers` to `models` to fit the style of the other services as
well as some of the internal files here.

In models.item I changed OzInBag to OZ in Bag since Go's naming convention
dictates that abbreviations should be all upper case or all lowercase, not
camel case. This should be changed in other model files, but I used this as
an example.

In models.item, I changed the json property names to be camel case not
underscore case to fit with general JSON styling as well as other service
styling. This also affected database names, and other models should be
changed to fit this convention.

In handlers, I updated both order and suborder to use their Helper rather
than calling SQL in the function itself. I also updated all the handlers
to use the base handler responses rather than directly calling ctx.JSON to
fit conventions with other services and to make json responses consistent
across services.

In router, I updated the routes to reflect actual objects and interfaces
in the latest version of the project.

It seems like quite of a few of these errors could be avoided by making sure
the project compiles successfully before pushing to master. In the future,
we should not have changes that don't compile being pushed to master branch.
Let me know if there you're having any trouble with the go compiler or other
go related commands. I've found [this sublime plugin](https://github.com/DisposaBoy/GoSublime) useful for go development.